# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: contact
  - name: profile
  - name: calculators
  - name: workout
  - name: diet
  - name: plans
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculators/compute:
    post:
      operationId: computeCalculators
      tags: [calculators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculatorInput"
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResults"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workout/generate:
    post:
      operationId: generateWorkoutPlan
      tags: [workout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutGenerateInput"
      responses:
        "200":
          description: Generated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedWorkoutPlan"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /diet/generate:
    post:
      operationId: generateDietPlan
      tags: [diet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DietGenerateInput"
      responses:
        "200":
          description: Generated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedDietPlan"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlansListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/workout/{id}:
    delete:
      operationId: deleteWorkoutPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/diet/{id}:
    delete:
      operationId: deleteDietPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # ── Progress ──────────────────────────────────────────────────────────────
  /progress:
    get:
      operationId: listProgressLogs
      tags: [progress]
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressLog"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProgressLog
      tags: [progress]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressLog"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/{id}:
    delete:
      operationId: deleteProgressLog
      tags: [progress]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # ── Personal Records ──────────────────────────────────────────────────────
  /progress/prs:
    get:
      operationId: listPrRecords
      tags: [progress]
      responses:
        "200":
          description: PR records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPrRecord
      tags: [progress]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrRecord"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/prs/{id}:
    delete:
      operationId: deletePrRecord
      tags: [progress]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ContactInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        subject:
          type: string
        message:
          type: string
          minLength: 1
        type:
          type: string
          enum: [contact, newsletter]
      required: [name, email, message]
    NewsletterInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
      required: [email]
    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        goal:
          type: ["string", "null"]
        activityLevel:
          type: ["string", "null"]
        workoutExperience:
          type: ["string", "null"]
        medicalConditions:
          type: ["string", "null"]
        injuries:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        armsCm:
          type: ["number", "null"]
        thighsCm:
          type: ["number", "null"]
        shouldersCm:
          type: ["number", "null"]
        dailyCalories:
          type: ["integer", "null"]
        foodPreference:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        stressLevel:
          type: ["integer", "null"]
        waterIntakeLiters:
          type: ["number", "null"]
        gymAvailability:
          type: ["boolean", "null"]
        equipment:
          type: ["string", "null"]
        workoutDaysPerWeek:
          type: ["integer", "null"]
        preferredWorkoutTime:
          type: ["string", "null"]
        onboardingCompleted:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, createdAt, updatedAt]
    ProfileInput:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        goal:
          type: ["string", "null"]
        activityLevel:
          type: ["string", "null"]
        workoutExperience:
          type: ["string", "null"]
        medicalConditions:
          type: ["string", "null"]
        injuries:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        armsCm:
          type: ["number", "null"]
        thighsCm:
          type: ["number", "null"]
        shouldersCm:
          type: ["number", "null"]
        dailyCalories:
          type: ["integer", "null"]
        foodPreference:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        stressLevel:
          type: ["integer", "null"]
        waterIntakeLiters:
          type: ["number", "null"]
        gymAvailability:
          type: ["boolean", "null"]
        equipment:
          type: ["string", "null"]
        workoutDaysPerWeek:
          type: ["integer", "null"]
        preferredWorkoutTime:
          type: ["string", "null"]
        onboardingCompleted:
          type: ["boolean", "null"]
    # ── Calculators ──────────────────────────────────────────────────────────
    CalculatorInput:
      type: object
      properties:
        weightKg:
          type: number
        heightCm:
          type: number
        age:
          type: integer
        gender:
          type: string
        activityLevel:
          type: string
        bodyFatPercent:
          type: ["number", "null"]
        wristCm:
          type: ["number", "null"]
        hipCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        forearmCm:
          type: ["number", "null"]
        oneRepMaxWeightKg:
          type: ["number", "null"]
        oneRepMaxReps:
          type: ["integer", "null"]
        exerciseMets:
          type: ["number", "null"]
        exerciseDurationMin:
          type: ["integer", "null"]
        runDistanceKm:
          type: ["number", "null"]
        runTimeMin:
          type: ["number", "null"]
      required: [weightKg, heightCm, age, gender, activityLevel]
    CalculatorResults:
      type: object
      properties:
        bmi:
          type: number
        bmiCategory:
          type: string
        bmr:
          type: number
        tdee:
          type: number
        fatLossCalories:
          type: number
        muscleGainCalories:
          type: number
        maintenanceCalories:
          type: number
        proteinGrams:
          type: number
        carbsGrams:
          type: number
        fatGrams:
          type: number
        waterIntakeLiters:
          type: number
        leanBodyMassKg:
          type: ["number", "null"]
        bodyFatKg:
          type: ["number", "null"]
        idealWeightKgLow:
          type: number
        idealWeightKgHigh:
          type: number
        oneRepMaxKg:
          type: ["number", "null"]
        caloriesBurned:
          type: ["number", "null"]
        pace:
          type: ["number", "null"]
      required:
        - bmi
        - bmiCategory
        - bmr
        - tdee
        - fatLossCalories
        - muscleGainCalories
        - maintenanceCalories
        - proteinGrams
        - carbsGrams
        - fatGrams
        - waterIntakeLiters
        - idealWeightKgLow
        - idealWeightKgHigh
    # ── Workout Generator ─────────────────────────────────────────────────────
    WorkoutGenerateInput:
      type: object
      properties:
        goal:
          type: string
        experience:
          type: string
        split:
          type: string
        daysPerWeek:
          type: integer
        location:
          type: string
        additionalNotes:
          type: ["string", "null"]
      required: [goal, experience, split, daysPerWeek, location]
    WorkoutExercise:
      type: object
      properties:
        name:
          type: string
        primaryMuscles:
          type: array
          items:
            type: string
        secondaryMuscles:
          type: array
          items:
            type: string
        sets:
          type: integer
        reps:
          type: string
        rest:
          type: string
        tempo:
          type: string
        rpe:
          type: integer
        tips:
          type: string
        commonMistakes:
          type: string
      required: [name, primaryMuscles, secondaryMuscles, sets, reps, rest, tempo, rpe, tips, commonMistakes]
    WorkoutDay:
      type: object
      properties:
        day:
          type: string
        sessionName:
          type: string
        targetMuscles:
          type: string
        estimatedDurationMin:
          type: integer
        warmup:
          type: string
        cooldown:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
      required: [day, sessionName, targetMuscles, estimatedDurationMin, warmup, cooldown, exercises]
    WorkoutPlanData:
      type: object
      properties:
        programName:
          type: string
        duration:
          type: string
        overview:
          type: string
        weeklySchedule:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutDay"
        progressionNotes:
          type: string
      required: [programName, duration, overview, weeklySchedule, progressionNotes]
    SavedWorkoutPlan:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        goal:
          type: string
        experience:
          type: string
        split:
          type: string
        daysPerWeek:
          type: integer
        location:
          type: string
        additionalNotes:
          type: ["string", "null"]
        plan:
          $ref: "#/components/schemas/WorkoutPlanData"
        createdAt:
          type: string
      required: [id, userId, goal, experience, split, daysPerWeek, location, plan, createdAt]
    # ── Diet Generator ────────────────────────────────────────────────────────
    DietGenerateInput:
      type: object
      properties:
        cuisine:
          type: string
        calorieTarget:
          type: integer
        dietStyle:
          type: string
        allergies:
          type: ["string", "null"]
      required: [cuisine, calorieTarget, dietStyle]
    DietMeal:
      type: object
      properties:
        name:
          type: string
        time:
          type: string
        items:
          type: array
          items:
            type: string
        calories:
          type: integer
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        fiberG:
          type: number
      required: [name, time, items, calories, proteinG, carbsG, fatG, fiberG]
    DietMacros:
      type: object
      properties:
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        fiberG:
          type: number
        sugarG:
          type: number
      required: [proteinG, carbsG, fatG, fiberG, sugarG]
    DietPlanData:
      type: object
      properties:
        dailyCalorieTarget:
          type: integer
        macros:
          $ref: "#/components/schemas/DietMacros"
        meals:
          type: array
          items:
            $ref: "#/components/schemas/DietMeal"
        shoppingList:
          type: array
          items:
            type: string
        estimatedWeeklyBudgetUSD:
          type: number
        nutritionTips:
          type: string
      required: [dailyCalorieTarget, macros, meals, shoppingList, estimatedWeeklyBudgetUSD, nutritionTips]
    SavedDietPlan:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        cuisine:
          type: string
        calorieTarget:
          type: integer
        dietStyle:
          type: string
        allergies:
          type: ["string", "null"]
        plan:
          $ref: "#/components/schemas/DietPlanData"
        createdAt:
          type: string
      required: [id, userId, cuisine, calorieTarget, dietStyle, plan, createdAt]
    # ── Plans List ────────────────────────────────────────────────────────────
    PlansListResponse:
      type: object
      properties:
        workoutPlans:
          type: array
          items:
            $ref: "#/components/schemas/SavedWorkoutPlan"
        dietPlans:
          type: array
          items:
            $ref: "#/components/schemas/SavedDietPlan"
      required: [workoutPlans, dietPlans]
    # ── Progress Logs ─────────────────────────────────────────────────────────
    ProgressLogInput:
      type: object
      properties:
        logDate:
          type: string
        weightKg:
          type: ["number", "null"]
        bodyFatPercent:
          type: ["number", "null"]
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        armsCm:
          type: ["number", "null"]
        thighsCm:
          type: ["number", "null"]
        shouldersCm:
          type: ["number", "null"]
        caloriesConsumed:
          type: ["integer", "null"]
        waterLiters:
          type: ["number", "null"]
        sleepHours:
          type: ["number", "null"]
        stressLevel:
          type: ["integer", "null"]
        workoutCompleted:
          type: ["boolean", "null"]
        workoutDurationMin:
          type: ["integer", "null"]
        workoutNotes:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
      required: [logDate]
    ProgressLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        logDate:
          type: string
        weightKg:
          type: ["number", "null"]
        bodyFatPercent:
          type: ["number", "null"]
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        armsCm:
          type: ["number", "null"]
        thighsCm:
          type: ["number", "null"]
        shouldersCm:
          type: ["number", "null"]
        caloriesConsumed:
          type: ["integer", "null"]
        waterLiters:
          type: ["number", "null"]
        sleepHours:
          type: ["number", "null"]
        stressLevel:
          type: ["integer", "null"]
        workoutCompleted:
          type: ["boolean", "null"]
        workoutDurationMin:
          type: ["integer", "null"]
        workoutNotes:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, logDate, createdAt, updatedAt]
    ProgressSummaryPeriod:
      type: object
      properties:
        avgWeightKg:
          type: ["number", "null"]
        avgBodyFatPercent:
          type: ["number", "null"]
        avgCaloriesConsumed:
          type: ["number", "null"]
        avgWaterLiters:
          type: ["number", "null"]
        avgSleepHours:
          type: ["number", "null"]
        workoutCount:
          type: integer
        totalLogs:
          type: integer
      required: [workoutCount, totalLogs]
    ProgressSummary:
      type: object
      properties:
        weekly:
          $ref: "#/components/schemas/ProgressSummaryPeriod"
        monthly:
          $ref: "#/components/schemas/ProgressSummaryPeriod"
        allTime:
          $ref: "#/components/schemas/ProgressSummaryPeriod"
        streak:
          type: integer
        totalLogs:
          type: integer
      required: [weekly, monthly, allTime, streak, totalLogs]
    # ── Personal Records ─────────────────────────────────────────────────────
    PrRecordInput:
      type: object
      properties:
        exercise:
          type: string
        weightKg:
          type: ["number", "null"]
        reps:
          type: ["integer", "null"]
        distanceKm:
          type: ["number", "null"]
        durationSec:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        achievedAt:
          type: string
      required: [exercise, achievedAt]
    PrRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        exercise:
          type: string
        weightKg:
          type: ["number", "null"]
        reps:
          type: ["integer", "null"]
        distanceKm:
          type: ["number", "null"]
        durationSec:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        achievedAt:
          type: string
        createdAt:
          type: string
      required: [id, userId, exercise, achievedAt, createdAt]