# lib/api-zod/src/generated/types/prRecord.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PrRecord {
  id: string;
  userId: string;
  exercise: string;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  durationSec?: number | null;
  /** @nullable */
  notes?: string | null;
  achievedAt: string;
  createdAt: string;
}